Add Context for fetching color definitions from API
