+++ src/main.jsx
@@ -2,9 +2,12 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './main.css'
 import App from './App.jsx'
+import { ColorDefinitionsProvider } from './context/colors.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <ColorDefinitionsProvider>
+      <App />
+    </ColorDefinitionsProvider>
   </StrictMode>,
 )
